requirements.txt
fastapi
python-telegram-bot==20.8
requests
python-dotenv
uvicorn
